+++ CHANGES.md
@@ -11,6 +11,10 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
+CHANGES  4.3.1.2 19/03/2025
+---------------------------
+    * change tables split in row termination
+
 CHANGES  4.3.1.1 14/03/2025
 ---------------------------
     * allow for missing flowable.height in Table.splitCell